requirements.txt
flask==2.3.2
flask-cors==4.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
gunicorn==20.1.0
logging==0.5.1.2
google-api-core==2.17.1
protobuf==4.25.3
